Reorganize packaged
COAND-860
