+++ example-app/src/main/java/com/adyen/checkout/example/ui/settings/model/SettingsCategory.kt
@@ -3,10 +3,10 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by josephj on 23/8/2024.
+ * Created by josephj on 3/9/2024.
  */
 
-package com.adyen.checkout.example.ui.settings
+package com.adyen.checkout.example.ui.settings.model
 
 import androidx.annotation.StringRes
